update for praktikum12

+++ database/migrations/2024_12_06_121456_provinsi.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('nama',45);
             $table->string('alt_nama',45);
-            $table->double('latitude');
-            $table->double('longitude');
+            $table->decimal('latitude',10,8);
+            $table->decimal('longitude',11,8);
             $table->timestamps();
         });
     }
